--- packages-private/diagnostics-shared/src/index.ts
@@ -1,9 +1,0 @@
-/*
- * SPDX-FileCopyrightText: syuilo and misskey-project
- * SPDX-License-Identifier: AGPL-3.0-only
- */
-
-export * from './env';
-export * from './format';
-export * from './html';
-export * from './stats';
